test(recovery): keep performance gates comparable
