+++ src/tests/mod.rs
@@ -44,6 +44,7 @@ mod support;
 mod upstream_reconciliation;
 mod upstream_reconciliation_continuation;
 mod upstream_reconciliation_fencing;
+mod upstream_reconciliation_projection;
 mod usage_series_and_backfills;
 mod user_business_calls_1h;
 mod user_pressure_analysis;
