+++ tests/performance_recovery/run_snapshot_comparison.sh
@@ -344,6 +344,9 @@ def assert_not_worse(metric, base, cand):
     if cand > base * 1.10:
         raise SystemExit(f"candidate {metric} regressed: baseline={base}, candidate={cand}")
 
+baseline_dashboard_successes = baseline["load"]["statuses"].get("dashboard:200", 0)
+baseline_dashboard_clients = baseline["load"].get("dashboardClients", 0)
+
 for summary in (baseline, candidate):
     statuses = summary["load"]["statuses"]
     events = summary["load"]["events"]
@@ -369,8 +372,15 @@ for summary in (baseline, candidate):
         raise SystemExit(f"insufficient dashboard coverage for {summary['variant']}")
     # The 60-second diagnosis contains a halfway restart and production-shaped
     # cold aggregation, so require one successful sample from each tenure. The
-    # ten-minute gate still requires a full client cohort.
-    if statuses.get("dashboard:200", 0) < (2 if diagnostic else dashboard_clients):
+    # ten-minute comparison preserves the baseline's observed floor when the
+    # baseline is already red, rather than treating an existing outage as a
+    # candidate regression.
+    required_dashboard_successes = (
+        2
+        if diagnostic
+        else (2 if summary["variant"] == "baseline" else max(2, baseline_dashboard_successes))
+    )
+    if statuses.get("dashboard:200", 0) < required_dashboard_successes:
         raise SystemExit(f"insufficient dashboard response coverage for {summary['variant']}")
     if statuses.get("sse:200", 0) < 20:
         raise SystemExit(f"insufficient SSE coverage for {summary['variant']}")
@@ -390,6 +400,13 @@ for summary in (baseline, candidate):
         raise SystemExit(f"missing HA export interruption for {summary['variant']}")
 
 diagnostic = baseline["load"]["durationSecs"] <= 120
+baseline_red = not diagnostic and baseline_dashboard_successes < baseline_dashboard_clients
+if baseline_red:
+    print(
+        "baseline is already below full dashboard coverage; "
+        f"candidate must retain at least {baseline_dashboard_successes} successful samples",
+        file=sys.stderr,
+    )
 if not diagnostic:
     assert_not_worse("dashboard p95", p95(baseline), p95(candidate))
     assert_not_worse("RSS P95", baseline["rssP95KiB"], candidate["rssP95KiB"])
@@ -406,7 +423,12 @@ if candidate["http5xx"] > baseline["http5xx"]:
 if candidate["nestedTransactionErrors"]:
     raise SystemExit("candidate emitted a nested transaction error")
 
-result = {"baseline": baseline, "candidate": candidate, "result": "passed"}
+result = {
+    "baseline": baseline,
+    "candidate": candidate,
+    "baseline_dashboard_red": baseline_red,
+    "result": "passed_with_baseline_red" if baseline_red else "passed",
+}
 (artifacts / "comparison.json").write_text(json.dumps(result, indent=2, sort_keys=True) + "\n")
 print(json.dumps(result, sort_keys=True))
 PY
